Updated Esign pricing
- Updated the cost of each token.
- Added cleanup for functions used across the app

Closes #344
