+++ netlify/functions/0029_SendPreparedEsignDocument.js
@@ -330,7 +330,7 @@ const processEmailService = async (data) => {
   );
 
   const response = await fetch(
-    `${process.env.SITE_URL}/.netlify/functions/0001_send_email_fn`,
+    `${process.env.SITE_URL}/.netlify/functions/0001_SendCustomEmail`,
     {
       method: "POST",
       headers: {
